+++ src/main/java/me/unariginal/novaraids/data/Location.java
@@ -3,7 +3,8 @@ package me.unariginal.novaraids.data;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.util.math.Vec3d;
 
-public record Location(String name,
+public record Location(String id,
+                       String name,
                        Vec3d pos,
                        ServerWorld world,
                        int border_radius,